internal: remove metrics.tarantool circular deps
Now all code related to registry initialization and manipulation is
extracted to a separate file metrics/api.lua. This patch doesn't change
any public API. This patch should resolve the circular dependency
between metrics and metrics.tarantool (including its dependencies).

Part of #433, part of tarantool/tarantool#7725
